# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_jherrma_caldav-server_internal_adapter_http_dto.CalDAVCredentialListResponse:
    properties:
      credentials:
        items:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.CalDAVCredentialResponse'
        type: array
    type: object
  github_com_jherrma_caldav-server_internal_adapter_http_dto.CalDAVCredentialResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      last_used_at:
        type: string
      last_used_ip:
        type: string
      name:
        type: string
      permission:
        type: string
      username:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateCalDAVCredentialRequest:
    properties:
      expires_at:
        type: string
      name:
        type: string
      password:
        type: string
      permission:
        type: string
      username:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_adapter_http_dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.UserResponse'
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.RegisterRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  github_com_jherrma_caldav-server_internal_adapter_http_dto.RegisterResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      is_active:
        type: boolean
      username:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_adapter_http_dto.UserResponse:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
...
  github_com_jherrma_caldav-server_internal_usecase_apppassword.CreateCalDAVCredentialOutput:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      name:
        type: string
      permission:
        type: string
      username:
        type: string
    type: object
...
  internal_adapter_http.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: error_code
        type: string
      message:
        example: Human readable message
        type: string
    type: object
  internal_adapter_http.SuccessResponseBody:
    description: Success response body
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: HTTP Basic Authentication for DAV endpoints
  title: CalDAV/CardDAV Server API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: User email
        in: body
        name: request
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_http.SuccessResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Request password reset
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Login with email and password
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token to logout
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_http.SuccessResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Logout user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          properties:
            refresh_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              expires_at:
                type: integer
              token_type:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "409":
          description: Email already registered
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Register a new user
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Set new password using reset token from email
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          properties:
            new_password:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_http.SuccessResponseBody'
        "400":
          description: Invalid token
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Reset password with token
      tags:
      - Authentication
  /auth/verify:
    get:
      consumes:
      - application/json
      description: Verify user email with token from verification email
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapter_http.SuccessResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      summary: Verify email address
      tags:
      - Authentication
  /caldav-credentials:
    get:
      description: List all CalDAV credentials for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.CalDAVCredentialListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: List CalDAV credentials
      tags:
      - Credentials
    post:
      consumes:
      - application/json
      description: Create a new app-specific password/credential for CalDAV access
      parameters:
      - description: Credential details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_jherrma_caldav-server_internal_adapter_http_dto.CreateCalDAVCredentialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jherrma_caldav-server_internal_usecase_apppassword.CreateCalDAVCredentialOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create CalDAV credential
      tags:
      - Credentials
  /caldav-credentials/{id}:
    delete:
      description: Revoke/Delete a CalDAV credential
      parameters:
      - description: Credential UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_adapter_http.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Revoke CalDAV credential
      tags:
      - Credentials
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...